Replace 'photon' with 'protocol'
- Replace all instances of 'photon' with 'protocol'
- Change path names from 'github.com/FirefoxUX/design-tokens' to 'github.com/mozilla/protocol-design-tokens'
