--- photon-components/README.md
@@ -1,3 +1,0 @@
-# Photon Components
-
-A collection of components that harmonizes with Firefox’s new Photon design.
