--- photon-icons/README.md
@@ -1,3 +1,0 @@
-# Photon Icons
-
-A collection of icons that harmonize with Firefox’s new Photon design.
