+++ protocol-colors/build.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const colors = require('./photon-colors.json');
+const colors = require('./protocol-colors.json');
 const metadata = require('./package.json');
 const colorArray = [];
 
@@ -53,7 +53,7 @@ function createColor(color, element, format) {
 
 const formats = {
   'android': {
-    'output': [`<?xml version="1.0" encoding="utf-8"?>\n\n${xmlLicense}\n<resources>\n    <!-- Photon Color Palette v${metadata.version} -->\n`],
+    'output': [`<?xml version="1.0" encoding="utf-8"?>\n\n${xmlLicense}\n<resources>\n    <!-- Protocol Color Palette v${metadata.version} -->\n`],
     'formatter': (color, variant, value, alpha) => {
       if (alpha != '100') {
         variant += `_a${alpha}`;
@@ -65,7 +65,7 @@ const formats = {
     'footer': '</resources>'
   },
   'css': {
-    'output': [`${jsLicense}\n/* Photon Colors CSS Variables v${metadata.version} */\n\n:root {\n`],
+    'output': [`${jsLicense}\n/* Protocol Colors CSS Variables v${metadata.version} */\n\n:root {\n`],
     'formatter': (color, variant, value, alpha) => {
       if (alpha == '100') {
         return `  --${color}-${variant}: ${value};\n`;
@@ -78,7 +78,7 @@ const formats = {
     'ext': 'css'
   },
   'gimp': {
-    'output': [`GIMP Palette\nName: Photon Colors\n${shLicense}\n# Photon Colors GPL Color Palette v${metadata.version}\n# ${metadata.homepage}\n\n`],
+    'output': [`GIMP Palette\nName: Protocol Colors\n${shLicense}\n# Protocol Colors GPL Color Palette v${metadata.version}\n# ${metadata.homepage}\n\n`],
     'formatter': (color, variant, value, alpha) => {
       if (alpha == '100') {
         color = color.charAt(0).toUpperCase() + color.slice(1);
@@ -89,7 +89,7 @@ const formats = {
     'ext': 'gpl'
   },
   'ios': {
-    'output': [`${jsLicense}\n/* Photon Colors iOS Variables v${metadata.version}\n   From ${metadata.homepage} */\n\nextension UIColor {\n    struct Photon {\n`],
+    'output': [`${jsLicense}\n/* Protocol Colors iOS Variables v${metadata.version}\n   From ${metadata.homepage} */\n\nextension UIColor {\n    struct Protocol {\n`],
     'formatter': (color, variant, value, alpha) => {
       color = color[0].toUpperCase() + color.substr(1);
       if (alpha != '100') {
@@ -104,7 +104,7 @@ const formats = {
     'footer': '  }\n}'
   },
   'js': {
-    'output': [`${jsLicense}\n/* Photon Colors JS Variables v${metadata.version} */\n\n`],
+    'output': [`${jsLicense}\n/* Protocol Colors JS Variables v${metadata.version} */\n\n`],
     'formatter': (color, variant, value, alpha) => {
       if (alpha != '100') {
         variant += `_A${alpha}`;
@@ -115,7 +115,7 @@ const formats = {
     'ext': 'js'
   },
   'less': {
-    'output': [`${jsLicense}\n/* Photon Colors Less Variables v${metadata.version} */\n\n`],
+    'output': [`${jsLicense}\n/* Protocol Colors Less Variables v${metadata.version} */\n\n`],
     'formatter': (color, variant, value, alpha) => {
       if (alpha == '100') {
         return `@${color}-${variant}: ${value};\n`;
@@ -127,7 +127,7 @@ const formats = {
     'ext': 'less'
   },
   'libreoffice': {
-    'output': [`<?xml version="1.0" encoding="UTF-8"?>\n${xmlLicense}\n<ooo:color-table\n  xmlns:office="urn:oasis:names:tc:opendocument:xmlns:office:1.0"\n  xmlns:draw="urn:oasis:names:tc:opendocument:xmlns:drawing:1.0"\n  xmlns:xlink="http://www.w3.org/1999/xlink"\n  xmlns:svg="http://www.w3.org/2000/svg"\n  xmlns:ooo="http://openoffice.org/2004/office">\n<!-- Photon Color Palette v${metadata.version} -->\n\n`],
+    'output': [`<?xml version="1.0" encoding="UTF-8"?>\n${xmlLicense}\n<ooo:color-table\n  xmlns:office="urn:oasis:names:tc:opendocument:xmlns:office:1.0"\n  xmlns:draw="urn:oasis:names:tc:opendocument:xmlns:drawing:1.0"\n  xmlns:xlink="http://www.w3.org/1999/xlink"\n  xmlns:svg="http://www.w3.org/2000/svg"\n  xmlns:ooo="http://openoffice.org/2004/office">\n<!-- Protocol Color Palette v${metadata.version} -->\n\n`],
     'formatter': (color, variant, value, alpha) => {
       if (alpha == '100') {
         return `  <draw:color draw:name="${color}-${variant}" draw:color="${value}" />\n`;
@@ -137,7 +137,7 @@ const formats = {
     'footer': '</ooo:color-table>'
   },
   'sass': {
-    'output': [`${jsLicense}\n/* Photon Colors SCSS Variables v${metadata.version} */\n\n`],
+    'output': [`${jsLicense}\n/* Protocol Colors SCSS Variables v${metadata.version} */\n\n`],
     'formatter': (color, variant, value, alpha) => {
       if (alpha == '100') {
         return `$${color}-${variant}: ${value};\n`;
@@ -190,7 +190,7 @@ for (let key in formats) {
   if (!out_func) {
     out_func = (data) => data.join('');
   }
-  fs.writeFile(`photon-colors.${format.ext}`, out_func(format.output), 'utf8', (err) => {
+  fs.writeFile(`protocol-colors.${format.ext}`, out_func(format.output), 'utf8', (err) => {
     if (err) throw err;
   });
 }
