+++ protocol-colors/protocol-colors.css
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-/* Photon Colors CSS Variables v3.0.1 */
+/* Protocol Colors CSS Variables v3.0.1 */
 
 :root {
   --magenta-50: #ff1ad9;
