+++ protocol-colors/protocol-colors.scss
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-/* Photon Colors SCSS Variables v3.0.1 */
+/* Protocol Colors SCSS Variables v3.0.1 */
 
 $magenta-50: #ff1ad9;
 $magenta-60: #ed00b5;
