+++ protocol-components/README.md
@@ -1,0 +1,3 @@
+# Protocol Components
+
+A collection of components that harmonizes with Firefox’s new Protocol design.
