+++ protocol-components/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "photon-components",
+  "name": "protocol-components",
   "version": "0.0.1",
   "description": "Component theme for Firefox and related properties.",
   "scripts": {
@@ -7,7 +7,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/FirefoxUX/design-tokens.git"
+    "url": "git+https://github.com/mozilla/protocol-design-tokens.git"
   },
   "keywords": [
     "firefox",
@@ -19,7 +19,7 @@
   "author": "Blake Winton <bwinton@latte.ca>",
   "license": "MPL-2.0",
   "bugs": {
-    "url": "https://github.com/FirefoxUX/design-tokens/issues?q=is%3Aissue+is%3Aopen+label%3A%22tokens%3A+components%22"
+    "url": "https://github.com/mozilla/protocol-design-tokens/issues?q=is%3Aissue+is%3Aopen+label%3A%22tokens%3A+components%22"
   },
-  "homepage": "https://github.com/FirefoxUX/design-tokens/tree/master/photon-components#readme"
+  "homepage": "https://github.com/mozilla/protocol-design-tokens/tree/master/protocol-components#readme"
 }
