+++ protocol-icons/README.md
@@ -1,0 +1,3 @@
+# Protocol Icons
+
+A collection of icons that harmonize with Firefox’s new Protocol design.
