+++ protocol-icons/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "photon-icons",
+  "name": "protocol-icons",
   "version": "0.0.1",
   "description": "Icons for Firefox and related properties.",
   "scripts": {
@@ -7,7 +7,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/FirefoxUX/design-tokens.git"
+    "url": "git+https://github.com/mozilla/protocol-design-tokens.git"
   },
   "keywords": [
     "firefox",
@@ -19,9 +19,9 @@
   "author": "Blake Winton <bwinton@latte.ca>",
   "license": "MPL-2.0",
   "bugs": {
-    "url": "https://github.com/FirefoxUX/design-tokens/issues?q=is%3Aissue+is%3Aopen+label%3A%22tokens%3A+icons%22"
+    "url": "https://github.com/mozilla/protocol-design-tokens/issues?q=is%3Aissue+is%3Aopen+label%3A%22tokens%3A+icons%22"
   },
-  "homepage": "https://github.com/FirefoxUX/design-tokens/tree/master/photon-icons#readme",
+  "homepage": "https://github.com/mozilla/protocol-design-tokens/tree/master/protocol-icons#readme",
   "dependencies": {
     "shelljs": "0.8.1"
   }
